requirements.txt
asgiref==3.7.2
Django==4.1.2
django-autoslug==1.9.8
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.4.1
packaging==23.2
Pillow==9.2.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
